[refactor] 패키지 구조를 변경하라
controller, service, repository(database) 크게 3가지로 구분합니다.
이외에 각종 설정파일은 common에 넣고 빈 등록은 config에 넣습니다.
db에서 사용 할 정보들은 entity에 담습니다.
