+++ src/main/java/io/hhplus/tdd/common/exception/ExceededBalanceException.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.exception;
+package io.hhplus.tdd.common.exception;
 
 public class ExceededBalanceException extends RuntimeException {
     public ExceededBalanceException(long id, long point, long amount) {
