+++ src/main/java/io/hhplus/tdd/common/response/ApiControllerAdvice.java
@@ -1,7 +1,7 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.common.response;
 
-import io.hhplus.tdd.exception.ExceededBalanceException;
-import io.hhplus.tdd.exception.InsufficientBalanceException;
+import io.hhplus.tdd.common.exception.ExceededBalanceException;
+import io.hhplus.tdd.common.exception.InsufficientBalanceException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
