+++ src/main/java/io/hhplus/tdd/common/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.common.response;
 
 public record ErrorResponse(
         String code,
