+++ src/main/java/io/hhplus/tdd/domain/point/PointHistory.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.domain.point;
 
 public record PointHistory(
         long id,
