+++ src/main/java/io/hhplus/tdd/domain/point/TransactionType.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.domain.point;
 
 
 /**
