+++ src/test/java/io/hhplus/tdd/point/PointServiceTest.java
@@ -2,9 +2,13 @@ package io.hhplus.tdd.point;
 
 import io.hhplus.tdd.database.PointHistoryTable;
 import io.hhplus.tdd.database.UserPointTable;
-import io.hhplus.tdd.exception.ExceededBalanceException;
-import io.hhplus.tdd.exception.InsufficientBalanceException;
-import io.hhplus.tdd.util.UserLock;
+import io.hhplus.tdd.common.exception.ExceededBalanceException;
+import io.hhplus.tdd.common.exception.InsufficientBalanceException;
+import io.hhplus.tdd.common.util.UserLock;
+import io.hhplus.tdd.domain.point.PointHistory;
+import io.hhplus.tdd.domain.point.TransactionType;
+import io.hhplus.tdd.domain.point.UserPoint;
+import io.hhplus.tdd.service.PointService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
